% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{robas_2_data}
\alias{robas_2_data}
\title{Brushing data from the ROBAS 2 Study}
\format{
One row per brushing session
\describe{
  \item{User}{Participant ID}
  \item{Day.in.Study}{Days since start of study}
  etc
}
}
\usage{
robas_2_data
}
\description{
Brushing data from the ROBAS 2 Study
}
\keyword{datasets}
